alx_travel_app/requirement.txt
Django
djangorestframework
django-environ
django-cors-headers
drf-yasg
PyMySQL
celery
pika